+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # 0.0.9
 
+- `ObjectTypeDefinition` nodes now store an `isStatic` property. When `true`, it
+  means it is only usable in JSON contexts. When `false` (= default) it means
+  the object is only usable in Live contexts.
+- Disallows object type definitions to be used in hybrid contexts. Given a type
+  definition of `type Foo {}`, then either all references to `Foo` must be
+  `'Foo'` or all references must be `'LiveObject<Foo>`, but no mixing is
+  allowed. (Consequence of now having `isStatic` at the object type definition
+  level.)
 - Fix: some ugly error messages making incorrect suggestions
 
 # 0.0.8